+++ src/data/tpmt.rs
@@ -69,7 +69,7 @@ pub struct TpmtPublic {
 }
 
 impl TpmSized for TpmtPublic {
-    const SIZE: usize = TPM_MAX_COMMAND_SIZE;
+    const SIZE: usize = TPM_MAX_COMMAND_SIZE as usize;
     fn len(&self) -> usize {
         self.object_type.len()
             + self.name_alg.len()
